requirements.txt
aiogram>=3.0.0
openai
python-dotenv
httpx==0.27.2
torch
torchvision
pillow